refactor: move splash screen logic into own directory
